backend/requirements.txt
<<<<<<< HEAD
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
=======
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
openai>=1.30.0
gitpython>=3.1.40
bandit>=1.7.7
pylint>=3.1.0
radon>=6.0.1
httpx>=0.27.0
pydantic>=2.6.0
sse-starlette>=2.0.0
>>>>>>> 2e70c2b (PRISM Guardian AI full system (local LLM + pipeline))
